+++ examples/baselines/ppo/ppo.py
@@ -1,25 +1,34 @@
-from collections import defaultdict
 import os
-import random
-import time
-from dataclasses import dataclass
-from typing import Optional
 
-import gymnasium as gym
-import numpy as np
-import torch
-import torch.nn as nn
-import torch.optim as optim
-import tyro
-from torch.distributions.normal import Normal
-from torch.utils.tensorboard import SummaryWriter
-
-# ManiSkill specific imports
-import mani_skill.envs
 from mani_skill.utils import gym_utils
 from mani_skill.utils.wrappers.flatten import FlattenActionSpaceWrapper
 from mani_skill.utils.wrappers.record import RecordEpisode
 from mani_skill.vector.wrappers.gymnasium import ManiSkillVectorEnv
+
+os.environ["TORCHDYNAMO_INLINE_INBUILT_NN_MODULES"] = "1"
+
+import math
+import os
+import random
+import time
+from collections import defaultdict
+from dataclasses import dataclass
+from typing import Optional, Tuple
+
+import gymnasium as gym
+import numpy as np
+import tensordict
+import torch
+import torch.nn as nn
+import torch.optim as optim
+import tqdm
+import tyro
+from torch.utils.tensorboard import SummaryWriter
+import wandb
+from tensordict import from_module
+from tensordict.nn import CudaGraphModule
+from torch.distributions.normal import Normal
+
 
 @dataclass
 class Args:
@@ -37,25 +46,27 @@ class Args:
     """the wandb's project name"""
     wandb_entity: Optional[str] = None
     """the entity (team) of wandb's project"""
+    wandb_group: str = "PPO"
+    """the group of the run for wandb"""
     capture_video: bool = True
     """whether to capture videos of the agent performances (check out `videos` folder)"""
-    save_model: bool = True
+    save_trajectory: bool = False
+    """whether to save trajectory data into the `videos` folder"""
+    save_model: bool = False
     """whether to save model into the `runs/{run_name}` folder"""
     evaluate: bool = False
     """if toggled, only runs evaluation with the given model checkpoint and saves the evaluation trajectories"""
     checkpoint: Optional[str] = None
     """path to a pretrained checkpoint file to start evaluation/training from"""
 
-    # Algorithm specific arguments
+    # Environment specific arguments
     env_id: str = "PickCube-v1"
     """the id of the environment"""
-    total_timesteps: int = 10000000
-    """total timesteps of the experiments"""
-    learning_rate: float = 3e-4
-    """the learning rate of the optimizer"""
+    env_vectorization: str = "gpu"
+    """the type of environment vectorization to use"""
     num_envs: int = 512
     """the number of parallel environments"""
-    num_eval_envs: int = 8
+    num_eval_envs: int = 16
     """the number of parallel evaluation environments"""
     partial_reset: bool = True
     """whether to let parallel environments reset upon termination instead of truncation"""
@@ -67,6 +78,16 @@ class Args:
     """how often to reconfigure the environment during training"""
     eval_reconfiguration_freq: Optional[int] = 1
     """for benchmarking purposes we want to reconfigure the eval environment each reset to ensure objects are randomized in some tasks"""
+    eval_freq: int = 25
+    """evaluation frequency in terms of iterations"""
+    save_train_video_freq: Optional[int] = None
+    """frequency to save training videos in terms of iterations"""
+
+    # Algorithm specific arguments
+    total_timesteps: int = 10000000
+    """total timesteps of the experiments"""
+    learning_rate: float = 3e-4
+    """the learning rate of the optimizer"""
     anneal_lr: bool = False
     """Toggle learning rate annealing for policy and value networks"""
     gamma: float = 0.8
@@ -93,12 +114,7 @@ class Args:
     """the target KL divergence threshold"""
     reward_scale: float = 1.0
     """Scale the reward by this factor"""
-    eval_freq: int = 25
-    """evaluation frequency in terms of iterations"""
-    save_train_video_freq: Optional[int] = None
-    """frequency to save training videos in terms of iterations"""
-    finite_horizon_gae: bool = True
-
+    finite_horizon_gae: bool = False
 
     # to be filled in runtime
     batch_size: int = 0
@@ -108,6 +124,12 @@ class Args:
     num_iterations: int = 0
     """the number of iterations (computed in runtime)"""
 
+    # Torch optimizations
+    compile: bool = False
+    """whether to use torch.compile."""
+    cudagraphs: bool = False
+    """whether to use cudagraphs on top of compile."""
+
 def layer_init(layer, std=np.sqrt(2), bias_const=0.0):
     torch.nn.init.orthogonal_(layer.weight, std)
     torch.nn.init.constant_(layer.bias, bias_const)
@@ -115,46 +137,39 @@ def layer_init(layer, std=np.sqrt(2), bias_const=0.0):
 
 
 class Agent(nn.Module):
-    def __init__(self, envs):
+    def __init__(self, n_obs, n_act, device=None):
         super().__init__()
         self.critic = nn.Sequential(
-            layer_init(nn.Linear(np.array(envs.single_observation_space.shape).prod(), 256)),
+            layer_init(nn.Linear(n_obs, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, 256)),
+            layer_init(nn.Linear(256, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, 256)),
+            layer_init(nn.Linear(256, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, 1)),
+            layer_init(nn.Linear(256, 1, device=device)),
         )
         self.actor_mean = nn.Sequential(
-            layer_init(nn.Linear(np.array(envs.single_observation_space.shape).prod(), 256)),
+            layer_init(nn.Linear(n_obs, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, 256)),
+            layer_init(nn.Linear(256, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, 256)),
+            layer_init(nn.Linear(256, 256, device=device)),
             nn.Tanh(),
-            layer_init(nn.Linear(256, np.prod(envs.single_action_space.shape)), std=0.01*np.sqrt(2)),
+            layer_init(nn.Linear(256, n_act, device=device), std=0.01*np.sqrt(2)),
         )
-        self.actor_logstd = nn.Parameter(torch.ones(1, np.prod(envs.single_action_space.shape)) * -0.5)
+        self.actor_logstd = nn.Parameter(torch.zeros(1, n_act, device=device))
 
     def get_value(self, x):
         return self.critic(x)
-    def get_action(self, x, deterministic=False):
-        action_mean = self.actor_mean(x)
-        if deterministic:
-            return action_mean
-        action_logstd = self.actor_logstd.expand_as(action_mean)
-        action_std = torch.exp(action_logstd)
-        probs = Normal(action_mean, action_std)
-        return probs.sample()
-    def get_action_and_value(self, x, action=None):
-        action_mean = self.actor_mean(x)
+
+    def get_action_and_value(self, obs, action=None):
+        action_mean = self.actor_mean(obs)
         action_logstd = self.actor_logstd.expand_as(action_mean)
         action_std = torch.exp(action_logstd)
         probs = Normal(action_mean, action_std)
         if action is None:
-            action = probs.sample()
-        return action, probs.log_prob(action).sum(1), probs.entropy().sum(1), self.critic(x)
+            action = action_mean + action_std * torch.randn_like(action_mean)
+        return action, probs.log_prob(action).sum(1), probs.entropy().sum(1), self.critic(obs)
 
 class Logger:
     def __init__(self, log_wandb=False, tensorboard: SummaryWriter = None) -> None:
@@ -167,17 +182,135 @@ class Logger:
     def close(self):
         self.writer.close()
 
+def gae(next_obs, next_done, container, final_values):
+    # bootstrap value if not done
+    next_value = get_value(next_obs).reshape(-1)
+    lastgaelam = 0
+    nextnonterminals = (~container["dones"]).float().unbind(0)
+    vals = container["vals"]
+    vals_unbind = vals.unbind(0)
+    rewards = container["rewards"].unbind(0)
+
+    advantages = []
+    nextnonterminal = (~next_done).float()
+    nextvalues = next_value
+    for t in range(args.num_steps - 1, -1, -1):
+        cur_val = vals_unbind[t]
+        # real_next_values = nextvalues * nextnonterminal
+        real_next_values = nextnonterminal * nextvalues + final_values[t] # t instead of t+1
+        delta = rewards[t] + args.gamma * real_next_values - cur_val
+        advantages.append(delta + args.gamma * args.gae_lambda * nextnonterminal * lastgaelam)
+        lastgaelam = advantages[-1]
+
+        nextnonterminal = nextnonterminals[t]
+        nextvalues = cur_val
+
+    advantages = container["advantages"] = torch.stack(list(reversed(advantages)))
+    container["returns"] = advantages + vals
+    return container
+
+
+def rollout(obs, done):
+    ts = []
+    final_values = torch.zeros((args.num_steps, args.num_envs), device=device)
+    for step in range(args.num_steps):
+        # ALGO LOGIC: action logic
+        action, logprob, _, value = policy(obs=obs)
+
+        # TRY NOT TO MODIFY: execute the game and log data.
+        next_obs, reward, next_done, infos = step_func(action)
+
+        if "final_info" in infos:
+            final_info = infos["final_info"]
+            done_mask = infos["_final_info"]
+            for k, v in final_info["episode"].items():
+                logger.add_scalar(f"train/{k}", v[done_mask].float().mean(), global_step)
+            with torch.no_grad():
+                final_values[step, torch.arange(args.num_envs, device=device)[done_mask]] = agent.get_value(infos["final_observation"][done_mask]).view(-1)
+
+        ts.append(
+            tensordict.TensorDict._new_unsafe(
+                obs=obs,
+                # cleanrl ppo examples associate the done with the previous obs (not the done resulting from action)
+                dones=done,
+                vals=value.flatten(),
+                actions=action,
+                logprobs=logprob,
+                rewards=reward,
+                batch_size=(args.num_envs,),
+            )
+        )
+        # NOTE (stao): change here for gpu env
+        obs = next_obs = next_obs
+        done = next_done
+    # NOTE (stao): need to do .to(device) i think? otherwise container.device is None, not sure if this affects anything
+    container = torch.stack(ts, 0).to(device)
+    return next_obs, done, container, final_values
+
+
+def update(obs, actions, logprobs, advantages, returns, vals):
+    optimizer.zero_grad()
+    _, newlogprob, entropy, newvalue = agent.get_action_and_value(obs, actions)
+    logratio = newlogprob - logprobs
+    ratio = logratio.exp()
+
+    with torch.no_grad():
+        # calculate approx_kl http://joschu.net/blog/kl-approx.html
+        old_approx_kl = (-logratio).mean()
+        approx_kl = ((ratio - 1) - logratio).mean()
+        clipfrac = ((ratio - 1.0).abs() > args.clip_coef).float().mean()
+
+    if args.norm_adv:
+        advantages = (advantages - advantages.mean()) / (advantages.std() + 1e-8)
+
+    # Policy loss
+    pg_loss1 = -advantages * ratio
+    pg_loss2 = -advantages * torch.clamp(ratio, 1 - args.clip_coef, 1 + args.clip_coef)
+    pg_loss = torch.max(pg_loss1, pg_loss2).mean()
+
+    # Value loss
+    newvalue = newvalue.view(-1)
+    if args.clip_vloss:
+        v_loss_unclipped = (newvalue - returns) ** 2
+        v_clipped = vals + torch.clamp(
+            newvalue - vals,
+            -args.clip_coef,
+            args.clip_coef,
+        )
+        v_loss_clipped = (v_clipped - returns) ** 2
+        v_loss_max = torch.max(v_loss_unclipped, v_loss_clipped)
+        v_loss = 0.5 * v_loss_max.mean()
+    else:
+        v_loss = 0.5 * ((newvalue - returns) ** 2).mean()
+
+    entropy_loss = entropy.mean()
+    loss = pg_loss - args.ent_coef * entropy_loss + v_loss * args.vf_coef
+
+    loss.backward()
+    gn = nn.utils.clip_grad_norm_(agent.parameters(), args.max_grad_norm)
+    optimizer.step()
+
+    return approx_kl, v_loss.detach(), pg_loss.detach(), entropy_loss.detach(), old_approx_kl, clipfrac, gn
+
+
+update = tensordict.nn.TensorDictModule(
+    update,
+    in_keys=["obs", "actions", "logprobs", "advantages", "returns", "vals"],
+    out_keys=["approx_kl", "v_loss", "pg_loss", "entropy_loss", "old_approx_kl", "clipfrac", "gn"],
+)
+
 if __name__ == "__main__":
     args = tyro.cli(Args)
-    args.batch_size = int(args.num_envs * args.num_steps)
-    args.minibatch_size = int(args.batch_size // args.num_minibatches)
+
+    batch_size = int(args.num_envs * args.num_steps)
+    args.minibatch_size = batch_size // args.num_minibatches
+    args.batch_size = args.num_minibatches * args.minibatch_size
     args.num_iterations = args.total_timesteps // args.batch_size
     if args.exp_name is None:
         args.exp_name = os.path.basename(__file__)[: -len(".py")]
         run_name = f"{args.env_id}__{args.exp_name}__{args.seed}__{int(time.time())}"
     else:
         run_name = args.exp_name
-
 
     # TRY NOT TO MODIFY: seeding
     random.seed(args.seed)
@@ -187,22 +320,22 @@ if __name__ == "__main__":
 
     device = torch.device("cuda" if torch.cuda.is_available() and args.cuda else "cpu")
 
-    # env setup
+    ####### Environment setup #######
     env_kwargs = dict(obs_mode="state", control_mode="pd_joint_delta_pos", render_mode="rgb_array", sim_backend="gpu")
     envs = gym.make(args.env_id, num_envs=args.num_envs if not args.evaluate else 1, reconfiguration_freq=args.reconfiguration_freq, **env_kwargs)
-    eval_envs = gym.make(args.env_id, num_envs=args.num_eval_envs, reconfiguration_freq=args.eval_reconfiguration_freq, **env_kwargs)
+    eval_envs = gym.make(args.env_id, num_envs=args.num_eval_envs, reconfiguration_freq=args.eval_reconfiguration_freq, human_render_camera_configs=dict(shader_pack="default"), **env_kwargs)
     if isinstance(envs.action_space, gym.spaces.Dict):
         envs = FlattenActionSpaceWrapper(envs)
         eval_envs = FlattenActionSpaceWrapper(eval_envs)
-    if args.capture_video:
+    if args.capture_video or args.save_trajectory:
         eval_output_dir = f"runs/{run_name}/videos"
         if args.evaluate:
             eval_output_dir = f"{os.path.dirname(args.checkpoint)}/test_videos"
-        print(f"Saving eval videos to {eval_output_dir}")
+        print(f"Saving eval trajectories/videos to {eval_output_dir}")
         if args.save_train_video_freq is not None:
             save_video_trigger = lambda x : (x // args.num_steps) % args.save_train_video_freq == 0
             envs = RecordEpisode(envs, output_dir=f"runs/{run_name}/train_videos", save_trajectory=False, save_video_trigger=save_video_trigger, max_steps_per_video=args.num_steps, video_fps=30)
-        eval_envs = RecordEpisode(eval_envs, output_dir=eval_output_dir, save_trajectory=args.evaluate, trajectory_name="trajectory", max_steps_per_video=args.num_eval_steps, video_fps=30)
+        eval_envs = RecordEpisode(eval_envs, output_dir=eval_output_dir, save_trajectory=args.save_trajectory, save_video=args.capture_video, trajectory_name="trajectory", max_steps_per_video=args.num_eval_steps, video_fps=30)
     envs = ManiSkillVectorEnv(envs, args.num_envs, ignore_terminations=not args.partial_reset, record_metrics=True)
     eval_envs = ManiSkillVectorEnv(eval_envs, args.num_eval_envs, ignore_terminations=True, record_metrics=True)
     assert isinstance(envs.single_action_space, gym.spaces.Box), "only continuous action space is supported"
@@ -223,8 +356,8 @@ if __name__ == "__main__":
                 config=config,
                 name=run_name,
                 save_code=True,
-                group="PPO",
-                tags=["ppo", "walltime_efficient"]
+                group=args.wandb_group,
+                tags=["ppo", "walltime_efficient", f"GPU:{torch.cuda.get_device_name()}"]
             )
         writer = SummaryWriter(f"runs/{run_name}")
         writer.add_text(
@@ -234,62 +367,85 @@ if __name__ == "__main__":
         logger = Logger(log_wandb=args.track, tensorboard=writer)
     else:
         print("Running evaluation")
+    n_act = math.prod(envs.single_action_space.shape)
+    n_obs = math.prod(envs.single_observation_space.shape)
+    assert isinstance(envs.single_action_space, gym.spaces.Box), "only continuous action space is supported"
 
-    agent = Agent(envs).to(device)
-    optimizer = optim.Adam(agent.parameters(), lr=args.learning_rate, eps=1e-5)
+    # Register step as a special op not to graph break
+    # @torch.library.custom_op("mylib::step", mutates_args=())
+    def step_func(action: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        # NOTE (stao): change here for gpu env
+        next_obs, reward, terminations, truncations, info = envs.step(action)
+        next_done = torch.logical_or(terminations, truncations)
+        return next_obs, reward, next_done, info
 
-    # ALGO Logic: Storage setup
-    obs = torch.zeros((args.num_steps, args.num_envs) + envs.single_observation_space.shape).to(device)
-    actions = torch.zeros((args.num_steps, args.num_envs) + envs.single_action_space.shape).to(device)
-    logprobs = torch.zeros((args.num_steps, args.num_envs)).to(device)
-    rewards = torch.zeros((args.num_steps, args.num_envs)).to(device)
-    dones = torch.zeros((args.num_steps, args.num_envs)).to(device)
-    values = torch.zeros((args.num_steps, args.num_envs)).to(device)
-
-    # TRY NOT TO MODIFY: start the game
-    global_step = 0
-    start_time = time.time()
-    next_obs, _ = envs.reset(seed=args.seed)
-    eval_obs, _ = eval_envs.reset(seed=args.seed)
-    next_done = torch.zeros(args.num_envs, device=device)
-    eps_returns = torch.zeros(args.num_envs, dtype=torch.float, device=device)
-    print(f"####")
-    print(f"args.num_iterations={args.num_iterations} args.num_envs={args.num_envs} args.num_eval_envs={args.num_eval_envs}")
-    print(f"args.minibatch_size={args.minibatch_size} args.batch_size={args.batch_size} args.update_epochs={args.update_epochs}")
-    print(f"####")
-    action_space_low, action_space_high = torch.from_numpy(envs.single_action_space.low).to(device), torch.from_numpy(envs.single_action_space.high).to(device)
-    def clip_action(action: torch.Tensor):
-        return torch.clamp(action.detach(), action_space_low, action_space_high)
-
+    ####### Agent #######
+    agent = Agent(n_obs, n_act, device=device)
     if args.checkpoint:
         agent.load_state_dict(torch.load(args.checkpoint))
+    # Make a version of agent with detached params
+    agent_inference = Agent(n_obs, n_act, device=device)
+    agent_inference_p = from_module(agent).data
+    agent_inference_p.to_module(agent_inference)
+
+    ####### Optimizer #######
+    optimizer = optim.Adam(
+        agent.parameters(),
+        lr=torch.tensor(args.learning_rate, device=device),
+        eps=1e-5,
+        capturable=args.cudagraphs and not args.compile,
+    )
+
+    ####### Executables #######
+    # Define networks: wrapping the policy in a TensorDictModule allows us to use CudaGraphModule
+    policy = agent_inference.get_action_and_value
+    get_value = agent_inference.get_value
+
+    # Compile policy
+    if args.compile:
+        policy = torch.compile(policy)
+        gae = torch.compile(gae, fullgraph=True)
+        update = torch.compile(update)
+
+    if args.cudagraphs:
+        policy = CudaGraphModule(policy)
+        gae = CudaGraphModule(gae)
+        update = CudaGraphModule(update)
+
+    global_step = 0
+    start_time = time.time()
+    container_local = None
+    next_obs = envs.reset()[0]
+    next_done = torch.zeros(args.num_envs, device=device, dtype=torch.bool)
+    pbar = tqdm.tqdm(range(1, args.num_iterations + 1))
 
     cumulative_times = defaultdict(float)
 
-    for iteration in range(1, args.num_iterations + 1):
-        print(f"Epoch: {iteration}, global_step={global_step}")
-        final_values = torch.zeros((args.num_steps, args.num_envs), device=device)
+    for iteration in pbar:
         agent.eval()
         if iteration % args.eval_freq == 1:
-            print("Evaluating")
             stime = time.perf_counter()
             eval_obs, _ = eval_envs.reset()
             eval_metrics = defaultdict(list)
             num_episodes = 0
             for _ in range(args.num_eval_steps):
                 with torch.no_grad():
-                    eval_obs, eval_rew, eval_terminations, eval_truncations, eval_infos = eval_envs.step(agent.get_action(eval_obs, deterministic=True))
+                    eval_obs, eval_rew, eval_terminations, eval_truncations, eval_infos = eval_envs.step(agent.actor_mean(eval_obs))
                     if "final_info" in eval_infos:
                         mask = eval_infos["_final_info"]
                         num_episodes += mask.sum()
                         for k, v in eval_infos["final_info"]["episode"].items():
                             eval_metrics[k].append(v)
-            print(f"Evaluated {args.num_eval_steps * args.num_eval_envs} steps resulting in {num_episodes} episodes")
+            eval_metrics_mean = {}
             for k, v in eval_metrics.items():
                 mean = torch.stack(v).float().mean()
+                eval_metrics_mean[k] = mean
                 if logger is not None:
                     logger.add_scalar(f"eval/{k}", mean, global_step)
-                print(f"eval_{k}_mean={mean}")
+            pbar.set_description(
+                f"success_once: {eval_metrics_mean['success_once']:.2f}, "
+                f"return: {eval_metrics_mean['return']:.2f}"
+            )
             if logger is not None:
                 eval_time = time.perf_counter() - stime
                 cumulative_times["eval_time"] += eval_time
@@ -304,159 +460,43 @@ if __name__ == "__main__":
         if args.anneal_lr:
             frac = 1.0 - (iteration - 1.0) / args.num_iterations
             lrnow = frac * args.learning_rate
-            optimizer.param_groups[0]["lr"] = lrnow
+            optimizer.param_groups[0]["lr"].copy_(lrnow)
 
+        torch.compiler.cudagraph_mark_step_begin()
         rollout_time = time.perf_counter()
-        for step in range(0, args.num_steps):
-            global_step += args.num_envs
-            obs[step] = next_obs
-            dones[step] = next_done
-
-            # ALGO LOGIC: action logic
-            with torch.no_grad():
-                action, logprob, _, value = agent.get_action_and_value(next_obs)
-                values[step] = value.flatten()
-            actions[step] = action
-            logprobs[step] = logprob
-
-            # TRY NOT TO MODIFY: execute the game and log data.
-            next_obs, reward, terminations, truncations, infos = envs.step(clip_action(action))
-            next_done = torch.logical_or(terminations, truncations).to(torch.float32)
-            rewards[step] = reward.view(-1) * args.reward_scale
-
-            if "final_info" in infos:
-                final_info = infos["final_info"]
-                done_mask = infos["_final_info"]
-                for k, v in final_info["episode"].items():
-                    logger.add_scalar(f"train/{k}", v[done_mask].float().mean(), global_step)
-                with torch.no_grad():
-                    final_values[step, torch.arange(args.num_envs, device=device)[done_mask]] = agent.get_value(infos["final_observation"][done_mask]).view(-1)
+        next_obs, next_done, container, final_values = rollout(next_obs, next_done)
         rollout_time = time.perf_counter() - rollout_time
         cumulative_times["rollout_time"] += rollout_time
-        # bootstrap value according to termination and truncation
-        with torch.no_grad():
-            next_value = agent.get_value(next_obs).reshape(1, -1)
-            advantages = torch.zeros_like(rewards).to(device)
-            lastgaelam = 0
-            for t in reversed(range(args.num_steps)):
-                if t == args.num_steps - 1:
-                    next_not_done = 1.0 - next_done
-                    nextvalues = next_value
-                else:
-                    next_not_done = 1.0 - dones[t + 1]
-                    nextvalues = values[t + 1]
-                real_next_values = next_not_done * nextvalues + final_values[t] # t instead of t+1
-                # next_not_done means nextvalues is computed from the correct next_obs
-                # if next_not_done is 1, final_values is always 0
-                # if next_not_done is 0, then use final_values, which is computed according to bootstrap_at_done
-                if args.finite_horizon_gae:
-                    """
-                    See GAE paper equation(16) line 1, we will compute the GAE based on this line only
-                    1             *(  -V(s_t)  + r_t                                                               + gamma * V(s_{t+1})   )
-                    lambda        *(  -V(s_t)  + r_t + gamma * r_{t+1}                                             + gamma^2 * V(s_{t+2}) )
-                    lambda^2      *(  -V(s_t)  + r_t + gamma * r_{t+1} + gamma^2 * r_{t+2}                         + ...                  )
-                    lambda^3      *(  -V(s_t)  + r_t + gamma * r_{t+1} + gamma^2 * r_{t+2} + gamma^3 * r_{t+3}
-                    We then normalize it by the sum of the lambda^i (instead of 1-lambda)
-                    """
-                    if t == args.num_steps - 1: # initialize
-                        lam_coef_sum = 0.
-                        reward_term_sum = 0. # the sum of the second term
-                        value_term_sum = 0. # the sum of the third term
-                    lam_coef_sum = lam_coef_sum * next_not_done
-                    reward_term_sum = reward_term_sum * next_not_done
-                    value_term_sum = value_term_sum * next_not_done
+        global_step += container.numel()
 
-                    lam_coef_sum = 1 + args.gae_lambda * lam_coef_sum
-                    reward_term_sum = args.gae_lambda * args.gamma * reward_term_sum + lam_coef_sum * rewards[t]
-                    value_term_sum = args.gae_lambda * args.gamma * value_term_sum + args.gamma * real_next_values
-
-                    advantages[t] = (reward_term_sum + value_term_sum) / lam_coef_sum - values[t]
-                else:
-                    delta = rewards[t] + args.gamma * real_next_values - values[t]
-                    advantages[t] = lastgaelam = delta + args.gamma * args.gae_lambda * next_not_done * lastgaelam # Here actually we should use next_not_terminated, but we don't have lastgamlam if terminated
-            returns = advantages + values
-
-        # flatten the batch
-        b_obs = obs.reshape((-1,) + envs.single_observation_space.shape)
-        b_logprobs = logprobs.reshape(-1)
-        b_actions = actions.reshape((-1,) + envs.single_action_space.shape)
-        b_advantages = advantages.reshape(-1)
-        b_returns = returns.reshape(-1)
-        b_values = values.reshape(-1)
+        update_time = time.perf_counter()
+        container = gae(next_obs, next_done, container, final_values)
+        container_flat = container.view(-1)
 
         # Optimizing the policy and value network
-        agent.train()
-        b_inds = np.arange(args.batch_size)
         clipfracs = []
-        update_time = time.perf_counter()
         for epoch in range(args.update_epochs):
-            np.random.shuffle(b_inds)
-            for start in range(0, args.batch_size, args.minibatch_size):
-                end = start + args.minibatch_size
-                mb_inds = b_inds[start:end]
+            b_inds = torch.randperm(container_flat.shape[0], device=device).split(args.minibatch_size)
+            for b in b_inds:
+                container_local = container_flat[b]
 
-                _, newlogprob, entropy, newvalue = agent.get_action_and_value(b_obs[mb_inds], b_actions[mb_inds])
-                logratio = newlogprob - b_logprobs[mb_inds]
-                ratio = logratio.exp()
-
-                with torch.no_grad():
-                    # calculate approx_kl http://joschu.net/blog/kl-approx.html
-                    old_approx_kl = (-logratio).mean()
-                    approx_kl = ((ratio - 1) - logratio).mean()
-                    clipfracs += [((ratio - 1.0).abs() > args.clip_coef).float().mean().item()]
-
-                if args.target_kl is not None and approx_kl > args.target_kl:
+                out = update(container_local, tensordict_out=tensordict.TensorDict())
+                clipfracs.append(out["clipfrac"])
+                if args.target_kl is not None and out["approx_kl"] > args.target_kl:
                     break
-
-                mb_advantages = b_advantages[mb_inds]
-                if args.norm_adv:
-                    mb_advantages = (mb_advantages - mb_advantages.mean()) / (mb_advantages.std() + 1e-8)
-
-                # Policy loss
-                pg_loss1 = -mb_advantages * ratio
-                pg_loss2 = -mb_advantages * torch.clamp(ratio, 1 - args.clip_coef, 1 + args.clip_coef)
-                pg_loss = torch.max(pg_loss1, pg_loss2).mean()
-
-                # Value loss
-                newvalue = newvalue.view(-1)
-                if args.clip_vloss:
-                    v_loss_unclipped = (newvalue - b_returns[mb_inds]) ** 2
-                    v_clipped = b_values[mb_inds] + torch.clamp(
-                        newvalue - b_values[mb_inds],
-                        -args.clip_coef,
-                        args.clip_coef,
-                    )
-                    v_loss_clipped = (v_clipped - b_returns[mb_inds]) ** 2
-                    v_loss_max = torch.max(v_loss_unclipped, v_loss_clipped)
-                    v_loss = 0.5 * v_loss_max.mean()
-                else:
-                    v_loss = 0.5 * ((newvalue - b_returns[mb_inds]) ** 2).mean()
-
-                entropy_loss = entropy.mean()
-                loss = pg_loss - args.ent_coef * entropy_loss + v_loss * args.vf_coef
-
-                optimizer.zero_grad()
-                loss.backward()
-                nn.utils.clip_grad_norm_(agent.parameters(), args.max_grad_norm)
-                optimizer.step()
-
-            if args.target_kl is not None and approx_kl > args.target_kl:
-                break
+            else:
+                continue
+            break
         update_time = time.perf_counter() - update_time
         cumulative_times["update_time"] += update_time
-        y_pred, y_true = b_values.cpu().numpy(), b_returns.cpu().numpy()
-        var_y = np.var(y_true)
-        explained_var = np.nan if var_y == 0 else 1 - np.var(y_true - y_pred) / var_y
 
         logger.add_scalar("charts/learning_rate", optimizer.param_groups[0]["lr"], global_step)
-        logger.add_scalar("losses/value_loss", v_loss.item(), global_step)
-        logger.add_scalar("losses/policy_loss", pg_loss.item(), global_step)
-        logger.add_scalar("losses/entropy", entropy_loss.item(), global_step)
-        logger.add_scalar("losses/old_approx_kl", old_approx_kl.item(), global_step)
-        logger.add_scalar("losses/approx_kl", approx_kl.item(), global_step)
-        logger.add_scalar("losses/clipfrac", np.mean(clipfracs), global_step)
-        logger.add_scalar("losses/explained_variance", explained_var, global_step)
-        print("SPS:", int(global_step / (time.time() - start_time)))
+        logger.add_scalar("losses/value_loss", out["v_loss"].item(), global_step)
+        logger.add_scalar("losses/policy_loss", out["pg_loss"].item(), global_step)
+        logger.add_scalar("losses/entropy", out["entropy_loss"].item(), global_step)
+        logger.add_scalar("losses/old_approx_kl", out["old_approx_kl"].item(), global_step)
+        logger.add_scalar("losses/approx_kl", out["approx_kl"].item(), global_step)
+        logger.add_scalar("losses/clipfrac", torch.stack(clipfracs).mean().cpu().item(), global_step)
         logger.add_scalar("charts/SPS", int(global_step / (time.time() - start_time)), global_step)
         logger.add_scalar("time/step", global_step, global_step)
         logger.add_scalar("time/update_time", update_time, global_step)
@@ -465,11 +505,6 @@ if __name__ == "__main__":
         for k, v in cumulative_times.items():
             logger.add_scalar(f"time/total_{k}", v, global_step)
         logger.add_scalar("time/total_rollout+update_time", cumulative_times["rollout_time"] + cumulative_times["update_time"], global_step)
-    if not args.evaluate:
-        if args.save_model:
-            model_path = f"runs/{run_name}/final_ckpt.pt"
-            torch.save(agent.state_dict(), model_path)
-            print(f"model saved to {model_path}")
-        logger.close()
+
     envs.close()
     eval_envs.close()
